+++ CHANGELOG.md
@@ -14,6 +14,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Removed
 
+## [1.1.0] - 2023-02-04
+
+### Fixed
+
+- Fixed codesmells in context adding the useMemo.
+- Fixed the cart context addItem and removeItem functions for having the correct logic.
+
 ## [1.0.0] - 2023-02-04
 
 ### Added